correction 9

+++ templates/propertys/controller.php
@@ -18,10 +18,9 @@ class Gm_Property_Controller extends Abstract_Template_Controller
 
     public function action() {
         if (isset($_GET['id'])) {
-            $param = sanitize_key($_GET['id']);
+            $param = (int) sanitize_key($_GET['id']);
         }
         $this->property_details = $this->wpdb->get_results("SELECT * FROM {$this->wpdb->prefix}gmt_property WHERE availability_id = 1 AND ID = {$param}");
-
         $this->render();
     }
